:bee: Minor fixes like spelling
In the Index.html, the "companies-container" div did not have a closing tag. Closing that allowed for the div to be centered on the page, and changing width from 100% to auto fixed the problem with the right arrow being on the other side of the page, and setting max width to 100% prevents it from going off the page... which would completely defeat the purpose of it.

I ran the Grunt file and that performed the changes to acm-1.0.1.min.js. I don't know what that changed. We also don't use the file so it doesn't really matter.